posts - complete

+++ app/controllers/Posts.php
@@ -151,6 +151,15 @@ class Posts extends Controller
     {
         if ($_SERVER['REQUEST_METHOD'] == 'POST')
         {
+            // Get existing post from model
+            $post = $this->postModel->getPostById($id);
+
+            //Check for owner
+            if ($post->user_id != $_SESSION['user_id'])
+            {
+                redirect('posts');
+            }
+
             if ($this->postModel->deletePost($id))
             {
                 flash('post_message', 'Post Removed');
